# backend/internal/launchcode/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Ant Browser Launch API
  description: |
    本地 LaunchServer HTTP API（仅 `127.0.0.1`）。

    ## 接口索引（共 14 个）

    | 分组 | 方法 | 路径 |
    |------|------|------|
    | 健康检查 | GET | `/api/health` |
    | 实例管理 | GET | `/api/profiles` |
    | 实例管理 | POST | `/api/profiles` |
    | 实例管理 | GET | `/api/profiles/{profileId}` |
    | 实例管理 | PUT | `/api/profiles/{profileId}` |
    | 实例管理 | DELETE | `/api/profiles/{profileId}` |
    | 实例管理 | GET | `/api/profiles/{profileId}/status` |
    | 实例管理 | POST | `/api/profiles/{profileId}/stop` |
    | 启动 | GET | `/api/launch/{code}` |
    | 启动 | POST | `/api/launch` |
    | 启动 | GET | `/api/launch/logs` |
    | 运行态 | POST | `/api/runtime/session` |
    | 运行态 | POST | `/api/runtime/status` |
    | 运行态 | POST | `/api/runtime/stop` |

    ## 认证

    若启用了 API Key，在 Authorize 中填写 header `X-Ant-Api-Key`。
    `GET /api/health` 不需要鉴权。

    ## 规则速查

    - `PUT /api/profiles/{id}` 是整份更新，不是 patch
    - 运行中的实例不能直接 `DELETE`，需先 stop
    - `matchMode=all` 仅 `POST /api/launch` 可用；runtime 接口仅支持 `unique` / `first`
    - `proxyId` 与 `proxyConfig` 同时传时优先 `proxyId`；`proxyId` 无效且无 `proxyConfig` 时返回 400
    - 启动/会话响应里的 `cdpUrl` 为实例调试端口直连地址（Launch Server 不转发 CDP）
  version: 1.0.0
servers:
  - url: /
    description: Current LaunchServer origin
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Ant-Api-Key
      description: 应用设置中启用 Launch API Key 后必填
  schemas:
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
        authHeader:
          type: string
          description: 鉴权失败时返回期望的 header 名
...
    LaunchRequestParams:
      type: object
      description: 仅影响本次启动的临时参数（不持久化覆盖实例配置）
      properties:
        launchArgs:
          type: array
          items:
            type: string
          description: 本次启动临时附加参数
        startUrls:
          type: array
          items:
            type: string
          description: 本次启动额外打开的网址
        skipDefaultStartUrls:
          type: boolean
          description: 是否跳过实例默认启动 URL
        proxyId:
          type: string
          description: 仅本次启动使用的代理池节点
        proxyConfig:
          type: string
          description: 仅本次启动使用的自定义代理链接
    LaunchSelector:
      type: object
      description: |
        目标实例选择条件。新接入推荐使用本对象；
        也可把同名字段放在请求体顶层（兼容写法）。
      properties:
        code:
          type: string
          description: launchCode
          example: BUYER_001
        key:
          type: string
          description: 兼容字段，语义同 code
        profileId:
          type: string
        profileName:
          type: string
        keyword:
          type: string
        keywords:
          type: array
          items:
            type: string
        tag:
          type: string
        tags:
          type: array
          items:
            type: string
        groupId:
          type: string
        matchMode:
          type: string
          enum: [unique, first, all]
          description: |
            unique=必须唯一命中；first=取第一个；all=批量（仅 POST /api/launch）
          example: unique
...
    LaunchRequest:
      allOf:
        - $ref: '#/components/schemas/LaunchRequestParams'
        - type: object
          properties:
            selector:
              $ref: '#/components/schemas/LaunchSelector'
            code:
              type: string
              description: 兼容写法，等价于 selector.code
            key:
              type: string
            profileId:
              type: string
            profileName:
              type: string
            keyword:
              type: string
            keywords:
              type: array
              items:
                type: string
            tag:
              type: string
            tags:
              type: array
              items:
                type: string
            groupId:
              type: string
            matchMode:
              type: string
              enum: [unique, first, all]
...
    LaunchSuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        profileId:
          type: string
        profileName:
          type: string
        launchCode:
          type: string
        pid:
          type: integer
        debugPort:
          type: integer
        debugReady:
          type: boolean
        runtimeWarning:
          type: string
        cdpPort:
          type: integer
          example: 9333
        cdpUrl:
          type: string
          example: http://127.0.0.1:9333
    LaunchBatchSuccessResponse:
      type: object
      description: matchMode=all 时的批量启动响应
      properties:
        ok:
          type: boolean
          example: true
        count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/LaunchSuccessResponse'
...
paths:
...
  /api/launch:
    post:
      tags: [Launch]
      operationId: launchBySelector
      summary: 按 selector 启动
      description: |
        更灵活的启动入口。支持 `selector` 或兼容顶层选择字段，以及 `launchArgs` / `startUrls` / `skipDefaultStartUrls`。

        - selector 为空且无任何兼容顶层选择字段时返回 400
        - `matchMode=all` 只在本接口可用，返回批量结果
        - `proxyId` / `proxyConfig` 只影响本次启动，不覆盖实例原代理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchRequest'
            examples:
              bySelector:
                summary: 按 keyword + tags
                value:
                  selector:
                    keyword: checkout
                    tags: [电商, 北美]
                    matchMode: unique
                  skipDefaultStartUrls: true
              byCodeCompat:
                summary: 顶层 code 兼容写法
                value:
                  code: BUYER_001
                  skipDefaultStartUrls: true
      responses:
        '200':
          description: 启动成功（单实例或批量）
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LaunchSuccessResponse'
                  - $ref: '#/components/schemas/LaunchBatchSuccessResponse'
        '400':
          description: selector 缺失或请求体非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未命中实例
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: selector 命中多个实例（unique 模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/launch/{code}:
    get:
      tags: [Launch]
      operationId: launchByCode
      summary: 按 launchCode 启动
      description: 最短路径启动接口，适合外部系统已拿到唯一 launchCode 的场景。
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          example: BUYER_001
      responses:
        '200':
          description: 启动成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchSuccessResponse'
              example:
                ok: true
                profileId: 550e8400-e29b-41d4-a716-446655440000
                profileName: buyer-001
                launchCode: BUYER_001
                pid: 10240
                debugPort: 9333
                debugReady: true
                cdpPort: 9333
                cdpUrl: http://127.0.0.1:9333
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: launchCode 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...